+++ gridcapa-core-valid-api/src/main/java/com/farao_community/farao/gridcapa_core_valid/api/exception/AbstractCoreValidException.java
@@ -1,11 +1,12 @@
 /*
- * Copyright (c) 2021, RTE (http://www.rte-france.com)
+ * Copyright (c) 2022, RTE (http://www.rte-france.com)
  *  This Source Code Form is subject to the terms of the Mozilla Public
  *  License, v. 2.0. If a copy of the MPL was not distributed with this
  *  file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ *
  */
-package com.farao_community.farao.core_valid.api.exception;
-import com.farao_community.farao.core_valid.api.JsonApiConverter;
+package com.farao_community.farao.gridcapa_core_valid.api.exception;
+import com.farao_community.farao.gridcapa_core_valid.api.JsonApiConverter;
 
 /**
  * Custom abstract exception to be extended by all application exceptions.
