+++ gridcapa-core-valid-api/src/main/java/com/farao_community/farao/gridcapa_core_valid/api/exception/CoreValidInternalException.java
@@ -1,25 +1,25 @@
 /*
- * Copyright (c) 2021, RTE (http://www.rte-france.com)
+ * Copyright (c) 2022, RTE (http://www.rte-france.com)
  *  This Source Code Form is subject to the terms of the Mozilla Public
  *  License, v. 2.0. If a copy of the MPL was not distributed with this
  *  file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ *
  */
 
-package com.farao_community.farao.core_valid.api.exception;
+package com.farao_community.farao.gridcapa_core_valid.api.exception;
 
 /**
  * @author Ameni Walha {@literal <ameni.walha at rte-france.com>}
  */
-public class CoreValidInvalidDataException extends AbstractCoreValidException {
+public class CoreValidInternalException extends AbstractCoreValidException {
+    private static final int STATUS = 500;
+    private static final String CODE = "500-InternalException";
 
-    private static final int STATUS = 400;
-    private static final String CODE = "400-InvalidDataException";
-
-    public CoreValidInvalidDataException(String message) {
+    public CoreValidInternalException(String message) {
         super(message);
     }
 
-    public CoreValidInvalidDataException(String message, Throwable throwable) {
+    public CoreValidInternalException(String message, Throwable throwable) {
         super(message, throwable);
     }
