+++ gridcapa-core-valid-api/src/test/java/com/farao_community/farao/gridcapa_core_valid/api/exception/CoreValidRaoExceptionTest.java
@@ -3,9 +3,10 @@
  *  This Source Code Form is subject to the terms of the Mozilla Public
  *  License, v. 2.0. If a copy of the MPL was not distributed with this
  *  file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ *
  */
 
-package com.farao_community.farao.core_valid.api.exception;
+package com.farao_community.farao.gridcapa_core_valid.api.exception;
 
 import org.junit.jupiter.api.Test;
