+++ gridcapa-core-valid-app/src/test/java/com/farao_community/farao/gridcapa_core_valid/app/services/UrlValidationServiceTest.java
@@ -7,7 +7,7 @@
 
 package com.farao_community.farao.gridcapa_core_valid.app.services;
 
-import com.farao_community.farao.core_valid.api.exception.CoreValidInvalidDataException;
+import com.farao_community.farao.gridcapa_core_valid.api.exception.CoreValidInvalidDataException;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
