+++ gridcapa-core-valid-spring-boot-starter/src/main/java/com/farao_community/farao/gridcapa_core_valid/starter/CoreValidResponseConversionHelper.java
@@ -6,9 +6,9 @@
  */
 package com.farao_community.farao.gridcapa_core_valid.starter;
 
-import com.farao_community.farao.core_valid.api.JsonApiConverter;
-import com.farao_community.farao.core_valid.api.exception.CoreValidInvalidDataException;
-import com.farao_community.farao.core_valid.api.resource.CoreValidResponse;
+import com.farao_community.farao.gridcapa_core_valid.api.JsonApiConverter;
+import com.farao_community.farao.gridcapa_core_valid.api.exception.CoreValidInvalidDataException;
+import com.farao_community.farao.gridcapa_core_valid.api.resource.CoreValidResponse;
 import com.github.jasminb.jsonapi.exceptions.ResourceParseException;
 import org.springframework.amqp.core.Message;
