+++ gridcapa-core-valid-spring-boot-starter/src/test/java/com/farao_community/farao/gridcapa_core_valid/starter/CoreValidResponseConversionHelperTest.java
@@ -7,9 +7,9 @@
 
 package com.farao_community.farao.gridcapa_core_valid.starter;
 
-import com.farao_community.farao.core_valid.api.JsonApiConverter;
-import com.farao_community.farao.core_valid.api.exception.CoreValidInvalidDataException;
-import com.farao_community.farao.core_valid.api.resource.CoreValidResponse;
+import com.farao_community.farao.gridcapa_core_valid.api.JsonApiConverter;
+import com.farao_community.farao.gridcapa_core_valid.api.exception.CoreValidInvalidDataException;
+import com.farao_community.farao.gridcapa_core_valid.api.resource.CoreValidResponse;
 import com.github.jasminb.jsonapi.exceptions.ResourceParseException;
 import com.github.jasminb.jsonapi.models.errors.Error;
 import com.github.jasminb.jsonapi.models.errors.Errors;
